refactor: moved rebuttal types to rebuttals folder
